新增日志记录

+++ takin-web-biz-service/src/main/java/io/shulie/takin/web/biz/pojo/request/webide/WebIDESyncScriptRequest.java
@@ -21,7 +21,7 @@ public class WebIDESyncScriptRequest {
 
     private String callbackAddr;
 
-    private Integer workRecordId;
+    private Long workRecordId;
 
 
     @Data
